+++ Twitter/TweetEditViewController.swift
@@ -13,19 +13,6 @@ protocol TweetEditViewControllerDelegate: AnyObject {
     func timeLineIconUpdate()
 }
 
-// データモデル
-class tweetDataModel: Object {
-    // 管理用 ID。プライマリーキー
-    @Persisted var id: String = UUID().uuidString //データを一意に識別するための識別子
-    // ツイート本文
-    @Persisted var tweetData: String = ""
-    // ユーザー名
-    @Persisted var userName: String = ""
-    // ユーザーアイコン
-    @Persisted var image: Data
-    
-}
-
 /// ツイート編集画面
 class TweetEditViewController: UIViewController {
